Add initial support for piscem-infer
This commit adds "piscem-infer" (perhaps we just want to call it "piscem")
to the list of supported methods.  This consists of reading in the
quantification files and the inferential replicates if available.

To read in the quantification files, we simply created the appropriate
reader setup. Piscem-infer uses a simple TSV format like the other tools
just with different column names.

To read in the inferential replicates, things are a *little* bit
different. Unlike sailfish, salmon, and kallisto, picem-infer doesn't
center its output in a directory-centric way.  Rather (based on
thoughtful feedback from users), it centers its output in a file-stem
based way. So, for example, if you quantify a sample with output stem
foo, you will have `foo.quant` and `foo.infreps.pq` if you choose to
compute inferential replicates.  The stem can be specified with a
directory if the user wants (e.g. `bar/foo`), but this doesn't change
the fact that the full stem will be the root for the multiple files
created.  This means that the inferential replicate reader for
`piscem-infer` works not by taking the directory for the quantification
output, but by taking the quantification file itself, stripping the
`.quant` prefix and appending a `.infreps.pq` prefix.  Otherwise, the
loading of the inferential replicates is very straightforward.

As one final note. Loading the inferential replicates requires reading
from the parquet format, for which the standard package seems to be
Apache's `arrow` package.  This will be required for loading infreps
from `piscem-infer`.  Should this be a recommended install in
`tximport`? If so, how should that be done?